+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Next
+* `Mutable::clone` is now lock-free.
+* Fixing bug with `from_stream` function.
 
 ## 0.3.22 - (2021-06-13)
 * Adding in `SignalExt::boxed` and `SignalExt::boxed_local` methods, for feature-parity with `FutureExt` and `StreamExt`.